requirements.txt
mss
pytesseract
Pillow
pyautogui